+++ gui/styles.py
@@ -1,13 +1,12 @@
 GUI_STYLES = """
 
 QLabel#apiKeyFieldLabel, QLabel#ipValueLabel, QLabel#statusHeadingLabel {
-    font-size: 15px;
+    font-size: 14px;
     font-weight: bold;
     color: #111827;
 }
 
 QLabel#ipValueField{
-    font-weight: bold;
     font-size: 14px;
     color: #111827;
 }
@@ -17,27 +16,16 @@ QLabel#statusHeadingLabel {
 }
 
 QLabel#apiKeyHelpTextLabel, QLabel#ipValueHelpTextLabel, QLabel#helpTextLabel, QLabel#appVersionLabel, QLabel#statusLabel {
-    font-size: 10px;
+    font-size: 11px;
     color: #6B7280;
-}
-
-QLabel#appVersionCheckLabel {
-    font-size: 10px;
-    color: #B45309;
 }
 
 QPushButton#startButton {
     background-color: #4338CA;
     color: white;
     font-size: 14px;
-    font-weight: bold;border: 0;
-}
-
-#loggerField {
-    background-color: transparent;
-    border: 1px solid #9CA3AF;
-    color: #4B5563;
-    font-size: 9px;
+    font-weight: bold;
+    border: 0;
 }
 
 """