+++ gui/window.py
@@ -1,5 +1,8 @@
-from PyQt5.QtWidgets import QWidget, QLabel, QPushButton, QLineEdit, QVBoxLayout, QFrame
+from PyQt5.QtWidgets import QWidget, QLabel, QPushButton, QLineEdit, \
+                            QVBoxLayout, QFrame
 from PyQt5.QtCore import Qt
+#from PyQt5.QtGui import QPixmap
+from PyQt5.QtSvg import QSvgWidget
 import logging
 import requests
 
@@ -8,6 +11,23 @@ from lib.file_handler import ConfigFile
 from receiver.receiver import RaceReceiver
 from receiver.helpers import get_local_ip
 import config
+
+
+class QHSeperationLine(QFrame):
+    """ Horizontal seperation line """
+    def __init__(self):
+        super().__init__()
+        self.setMinimumWidth(1)
+        self.setFixedHeight(20)
+        self.setFrameShape(QFrame.HLine)
+        self.setFrameShadow(QFrame.Sunken)
+
+class QVSpacer(QLabel):
+    """ Vertical empty space """
+    def __init__(self, height):
+        super().__init__()
+        self.setText(" ")
+        self.setFixedHeight(height)
 
 
 class MainWindow(QWidget):
@@ -34,6 +54,10 @@ class MainWindow(QWidget):
 
 
     def init_ui(self):
+        # 1) Logo & heading
+        logo_label = QSvgWidget('logo.svg')
+        logo_label.setFixedSize(100, 28)
+
         # 1) Enter API key section
         api_key_field_label = QLabel()
         api_key_field_label.setText("1) Enter your API key")
@@ -50,14 +74,13 @@ class MainWindow(QWidget):
         ip_value_label = QLabel()
         ip_value_label.setText("2) Check your F1 game Telemetry IP setting")
         ip_value_label.setObjectName("ipValueLabel")
-        ip_value_label.setContentsMargins(0, 20, 0, 0)
         ip_value_help_text_label = QLabel()
-        ip_value_help_text_label.setText("You can ignore this step if you're running this program on the same computer as the F1 game. Otherwise, open Settings --> Telemetry in the F1 game, and make sure the IP setting is set to this value.")
+        ip_value_help_text_label.setText("You can ignore this step if you're running this program on the same computer as the F1 game. Otherwise open Settings -> Telemetry in the F1 game and set the IP to this value.")
         ip_value_help_text_label.setObjectName("ipValueHelpTextLabel")
         ip_value_help_text_label.setWordWrap(True)
         self.ip_value = QLabel()
         self.ip_value.setObjectName("ipValueField")
-        self.ip_value.setContentsMargins(0, 5, 0, 20)
+        self.ip_value.setContentsMargins(0, 5, 0, 0)
 
         # Start/Stop button
         self.start_button = QPushButton('Start Telemetry')
@@ -73,37 +96,45 @@ class MainWindow(QWidget):
         help_text_label.setText("Need help? <a href='https://www.notion.so/F1Laps-Telemetry-Documentation-55ad605471624066aa67bdd45543eaf7'>Check out the Documentation & Help Center!</a>")
         help_text_label.setObjectName("helpTextLabel")
         help_text_label.setOpenExternalLinks(True)
-        help_text_label.setContentsMargins(0, 35, 0, 0)
-        app_version_label = QLabel()
-        app_version_label.setText("You're using F1Laps Telemetry version %s" % self.app_version)
-        app_version_label.setObjectName("appVersionLabel")
-        self.check_app_version_label = QLabel()
-        self.check_app_version_label.setObjectName("appVersionCheckLabel")
-        self.check_app_version_label.setOpenExternalLinks(True)
+        self.app_version_label = QLabel()
+        self.app_version_label.setText("You're using app version %s." % self.app_version)
+        self.app_version_label.setObjectName("appVersionLabel")
+        self.app_version_label.setWordWrap(True)
 
         # Draw layout
         self.layout = QVBoxLayout()
 
-        # API key section
+        # Logo section
+        self.layout.addWidget(logo_label, alignment=Qt.AlignCenter)
+        self.layout.addWidget(QHSeperationLine())
+
+        # API key & IP section
+        self.layout.addWidget(QVSpacer(0))
         self.layout.addWidget(api_key_field_label)
         self.layout.addWidget(api_key_help_text_label)
         self.layout.addWidget(self.api_key_field)
 
-        # IP section
+        self.layout.addWidget(QVSpacer(10))
         self.layout.addWidget(ip_value_label)
         self.layout.addWidget(ip_value_help_text_label)
         self.layout.addWidget(self.ip_value)
 
         # Start button
+        self.layout.addWidget(QVSpacer(2))
+        self.layout.addWidget(QHSeperationLine())
+        self.layout.addWidget(QVSpacer(5))
         self.layout.addWidget(self.start_button, alignment=Qt.AlignCenter)
         self.layout.addWidget(self.status_label, alignment=Qt.AlignCenter)
 
         # Status & help
+        self.layout.addWidget(QVSpacer(3))
+        self.layout.addWidget(QHSeperationLine())
         self.layout.addWidget(help_text_label)
-        self.layout.addWidget(app_version_label)
-        self.layout.setContentsMargins(30, 30, 30, 30)
+        self.layout.addWidget(self.app_version_label)
+        self.layout.setContentsMargins(30, 20, 30, 30)
         
         self.setLayout(self.layout)
+        self.setFixedWidth(420)
         self.setWindowTitle("F1Laps Telemetry") 
         log.info("Welcome to F1Laps Telemetry! You will see all logging in this text field.")
 
@@ -125,13 +156,12 @@ class MainWindow(QWidget):
             version = response.json()['version']
             user_version_int = int(self.app_version.replace(".", ""))
             current_version_int = int(version.replace(".", ""))
-            if version >= self.app_version:
-                self.check_app_version_label.setText("There's a new version available (%s).&nbsp;<a href='https://www.f1laps.com/api/telemetry_apps'>Download new version now!</a>" % version)
-                self.layout.addWidget(self.check_app_version_label)
-            elif version <= self.app_version:
-                self.check_app_version_label.setText("This is a pre-release version (stable version is %s)." % version)
-                self.check_app_version_label.setStyleSheet("color: #059669;")
-                self.layout.addWidget(self.check_app_version_label)
+            if version > self.app_version:
+                self.app_version_label.setText("There's a new app version available (you're on v%s).<br><a href='https://www.f1laps.com/api/telemetry_apps'>Click here to download the new version!</a>" % self.app_version)
+                self.app_version_label.setStyleSheet("color: #B45309")
+            elif version < self.app_version:
+                self.app_version_label.setText("This is pre-release version v%s (stable version is v%s)." % (self.app_version, version))
+                self.app_version_label.setStyleSheet("color: #059669")
         except Exception as ex:
             log.warning("Couldn't get most recent version from F1Laps due to: %s" % ex)